# openapi.yaml
openapi: 3.1.0
info:
  title: DreamWright API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamWright Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      operationId: listProjects
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Projects]
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: "My Webtoon"
              format: "webtoon"
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/status:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Projects]
      summary: Get project generation status
      description: Returns detailed status of all assets and generation progress
      operationId: getProjectStatus
      responses:
        '200':
          description: Project status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

  # Story endpoints - Resource-oriented: POST creates/expands the story
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ProjectFormat:
      type: string
      enum: [webtoon, short_drama]
    ProjectStatus:
      type: string
      enum: [draft, in_progress, completed]
    Genre:
      type: string
      enum: [romance, action, fantasy, thriller, slice_of_life, horror, comedy, drama, mystery, scifi]
    Tone:
      type: string
      enum: [comedic, dramatic, dark, lighthearted, romantic, suspenseful]
    CharacterRole:
      type: string
      enum: [protagonist, antagonist, supporting, minor]
    LocationType:
      type: string
      enum: [interior, exterior]
...
    ChapterStatus:
      type: string
      enum: [outlined, generating, completed]
...
    ErrorCode:
      type: string
      enum:
        - bad_request
        - validation_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - dependency_not_met
        - asset_exists
        - rate_limited
        - internal_error
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Paginated lists
    PaginatedProjectList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    ProjectSummary:
      type: object
      required: [id, name, format, status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/ProjectFormat'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        story_title:
          type: string
          nullable: true
        character_count:
          type: integer
        location_count:
          type: integer
        chapter_count:
          type: integer
    Project:
      type: object
      required: [id, name, format, status, created_at, characters, locations, chapters]
      properties:
        id:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/ProjectFormat'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        story:
          $ref: '#/components/schemas/Story'
          nullable: true
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterSummary'
        original_prompt:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        format:
          $ref: '#/components/schemas/ProjectFormat'
          default: webtoon
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProjectStatus'
    ProjectStatusResponse:
      type: object
      required: [project_id, status, story_expanded, characters, locations, chapters]
      properties:
        project_id:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        story_expanded:
          type: boolean
        characters:
          type: object
          required: [total, with_assets, missing_assets]
          properties:
            total:
              type: integer
            with_assets:
              type: integer
            missing_assets:
              type: array
              items:
                $ref: '#/components/schemas/AssetStatusItem'
        locations:
          type: object
          required: [total, with_assets, missing_assets]
          properties:
            total:
              type: integer
            with_assets:
              type: integer
            missing_assets:
              type: array
              items:
                $ref: '#/components/schemas/AssetStatusItem'
        chapters:
          type: object
          required: [total_beats, generated, with_panels]
          properties:
            total_beats:
              type: integer
            generated:
              type: integer
            with_panels:
              type: integer
            missing_chapters:
              type: array
              items:
                type: object
                required: [beat_number, beat_name]
                properties:
                  beat_number:
                    type: integer
                  beat_name:
                    type: string
        panels:
          type: object
          required: [total, generated]
          properties:
            total:
              type: integer
            generated:
              type: integer
    AssetStatusItem:
      type: object
      required: [id, name, asset_type]
      properties:
        id:
          type: string
        name:
          type: string
        asset_type:
          type: string
          enum: [portrait, reference, panel]

    # Story schemas
    StoryBeat:
      type: object
      required: [beat, description]
      properties:
        beat:
          type: string
        description:
          type: string
    Story:
      type: object
      required: [id, title, genre, tone, episode_count]
      properties:
        id:
          type: string
        title:
          type: string
        logline:
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        tone:
          $ref: '#/components/schemas/Tone'
        themes:
          type: array
          items:
            type: string
        target_audience:
          type: string
        episode_count:
          type: integer
        synopsis:
          type: string
        story_beats:
          type: array
          items:
            $ref: '#/components/schemas/StoryBeat'
...
    CharacterDescription:
      type: object
      properties:
        physical:
          type: string
        personality:
          type: string
        background:
          type: string
        motivation:
          type: string
    CharacterAssets:
      type: object
      properties:
        portrait:
          type: string
          nullable: true
          description: Path to portrait image
        portrait_url:
          type: string
          format: uri
          nullable: true
          description: URL to download portrait
        three_view:
          type: object
          properties:
            front:
              type: string
              nullable: true
            side:
              type: string
              nullable: true
            back:
              type: string
              nullable: true
    Character:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/CharacterAssets'
...
    LocationAssets:
      type: object
      properties:
        reference:
          type: string
          nullable: true
          description: Path to reference image
        reference_url:
          type: string
          format: uri
          nullable: true
          description: URL to download reference
    Location:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/LocationType'
        description:
          type: string
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/LocationAssets'
...
    ChapterSummary:
      type: object
      required: [id, number, title, status]
      properties:
        id:
          type: string
        number:
          type: integer
        title:
          type: string
        status:
          $ref: '#/components/schemas/ChapterStatus'
        scene_count:
          type: integer
        panel_count:
          type: integer
        generated_panel_count:
          type: integer
...